+++ Forest.h
@@ -91,7 +91,7 @@ public:
 	///
 	/// @exception FastCodeMLFatal Invalid range from command line
 	///
-	bool getBranchRange(const CmdLine& aCmdLine, size_t& aBranchStart, size_t& aBranchEnd, std::set<size_t>& aFgSet) const;
+	bool getBranchRange(const CmdLine& aCmdLine, size_t& aBranchStart, size_t& aBranchEnd, std::set<int>& aFgSet) const;
 
 	/// Reduce common subtrees on the whole forest.
 	///
@@ -180,7 +180,7 @@ public:
 	///
 	/// @return The internal branch index of the branches marked in the tree file. UINT_MAX otherwise.
 	///
-	std::set<size_t> getMarkedBranches(void) const {return mMarkedBranches;}
+	std::set<int> getMarkedBranches(void) const {return mMarkedBranches;}
 
 	/// Get site multiplicity values.
 	///
@@ -341,7 +341,7 @@ private:
 							mNodeNames;					///< List of node names. Zero is the root, then its first child and so on
 	std::vector<double>		mBranchLengths;				///< List of branch lengths (read from file or stored here to be exported in the tree file)
 	size_t					mMarkedInternalBranch;		///< Number of the internal branch as marked in the tree file
-	std::set<size_t>		mMarkedBranches;		    ///< Number of the marked branches in the tree file
+	std::set<int>		mMarkedBranches;		    ///< Number of the marked branches in the tree file
 
 #ifdef NEW_LIKELIHOOD
